% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanTopas-functions-001.R
\name{read.topas.demo}
\alias{read.topas.demo}
\title{read.topas.demo}
\usage{
read.topas.demo()
}
\value{
Text message
}
\description{
Demonstration of the read.topas functions
}
\details{
This function contains demonstration of the functions in clanTopas.
See function body.
}
