% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanTopas-functions-001.R
\name{read.topas.phasespace}
\alias{read.topas.phasespace}
\title{read.topas.phasespace}
\usage{
read.topas.phasespace(
  pn.full = "~//",
  fn.main = "",
  fn.scorer = "DoseScorer1",
  what = "Dose",
  what2 = "Dose",
  get.col.names.from.header = TRUE,
  number.of.columns = 28,
  include.meta.data = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{pn.full}{= Path to folder where scorer can be found}

\item{fn.main}{= First part of scorer filename.}

\item{fn.scorer}{= Last part of scorer filename (typically the name of the scorer).}

\item{what}{= Meta data like "Fluence of electrons".}

\item{what2}{= Meta data (supplement to what).}

\item{get.col.names.from.header}{= If TRUE, extract names of fields from file header.}

\item{number.of.columns}{= If get.col.names.from.header == FALSE, you will need to define the number of fields in the file.}

\item{include.meta.data}{=If TRUE, include the meta data (what, what2 + filename info) in returned dataframe.}

\item{verbose}{= TRUE or FALSE,}
}
\value{
a dataframe with scorer data included meta data, if requested.
}
\description{
Reading Topas phasespace file (header+phsp)
}
\details{
This function can extract the main information from the header file.
}
