% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanTopas-functions-001.R
\name{read.topas.simple}
\alias{read.topas.simple}
\title{read.topas.simple}
\usage{
read.topas.simple(...)
}
\arguments{
\item{pn.full}{= Path to folder where scorer can be found}

\item{fn.main}{= First part of scorer filename.}

\item{fn.scorer}{= Last part of scorer filename (typically the name of the scorer).}

\item{what}{= Meta data like "Fluence of electrons".}

\item{what2}{= Meta data (supplement to what).}

\item{get.col.names.from.header}{= If TRUE, extract names of fields from file header.}

\item{scorer.names.default}{= If get.col.names.from.header == FALSE, you will need to provide field names (e.g. c("Count_in_Bin","Mean","Sum"),)}

\item{include.meta.data}{=If TRUE, include the meta data (what, what2 + filename info) in returned dataframe.}

\item{verbose}{= TRUE or FALSE,}
}
\value{
a dataframe with scorer data included meta data, if requested.
}
\description{
Reading simple Topas scorer file (i.e. no xyz binning etc.)
}
\details{
This function can use the main information from the header file.
}
