% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanTopas-functions-001.R
\name{read.topas.spectrum}
\alias{read.topas.spectrum}
\title{read.topas.spectrum}
\usage{
read.topas.spectrum(
  pn.full = "~//",
  fn.main = "",
  fn.scorer = "DoseScorer1",
  what = "Dose",
  what2 = "Dose",
  get.col.names.from.header = TRUE,
  scorer.names.default = c("Count_in_Bin", "Mean", "Sum"),
  include.meta.data = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{MeV}{= electron kinetic energy}

\item{dat}{= list with parameters (both input and output)}

\item{dat$Z}{= atomic number}

\item{dat$A}{= atomic mass}

\item{dat$I}{= mean excitation energy in eV}

\item{dat$rho.density}{= density in g/cm3 (only used for the delta computation)}

\item{dat$fvec}{= sub-shell occupancy level (used in computation)}

\item{dat$Evec}{= sub-shell binding energy (used in computation)}

\item{dat$plot.wanted}{= TRUE or FALSE,}

\item{mu.solver.parm}{= search parameters for the mu.st solver}

\item{L.solver.parm}{= search parameters for the L solver}

\item{dat$nlev}{= number of sub-shells (output)}

\item{dat$exact.delta}{= the computed delta (output)}
}
\value{
a list (dat)
}
\description{
Reading Topas spectrum file (e.g. fluence spectra)
}
\details{
This function can use the main information from the header file.
}
